+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- Added HelpOptionHandler interface so that clients can have more control over the Sign In Help
+screen options. Core also provides default implementation.
+
+```kotlin
+@Provides
+    @Singleton
+    fun provideHelpOptionHandler(): HelpOptionHandler = DefaultHelpOptionHandler()
+```
 
 ### Changes
 - Challenge entities improved. Clients should update their DB with the following migration: